Fix tests after refactor
