+++ test/e2e_chat_completions_proxy_test.go
@@ -18,7 +18,7 @@ import (
 	"github.com/gaspardpetit/nfrx/internal/plugin"
 	"github.com/gaspardpetit/nfrx/internal/server"
 	"github.com/gaspardpetit/nfrx/internal/serverstate"
-	"github.com/gaspardpetit/nfrx/modules/llm/agent/internal/worker"
+	"github.com/gaspardpetit/nfrx/modules/llm/agent/worker"
 	llmext "github.com/gaspardpetit/nfrx/modules/llm/ext"
 	mcpext "github.com/gaspardpetit/nfrx/modules/mcp/ext"
 )
@@ -68,7 +68,7 @@ func TestE2EChatCompletionsProxy(t *testing.T) {
 	defer cancel()
 	wsURL := strings.Replace(srv.URL, "http", "ws", 1) + "/api/workers/connect"
 	go func() {
-		_ = worker.Run(ctx, config.WorkerConfig{ServerURL: wsURL, ClientKey: "secret", CompletionBaseURL: ollama.URL + "/v1", CompletionAPIKey: "secret-123", ClientID: "w1", ClientName: "w1", MaxConcurrency: 2, EmbeddingBatchSize: 0})
+		_ = worker.Run(ctx, worker.Config{ServerURL: wsURL, ClientKey: "secret", CompletionBaseURL: ollama.URL + "/v1", CompletionAPIKey: "secret-123", ClientID: "w1", ClientName: "w1", MaxConcurrency: 2, EmbeddingBatchSize: 0})
 	}()
 
 	// wait for worker registration
